requirements.txt
alembic==1.3.0
aniso8601==8.0.0
antiorm==1.2.1
blinker==1.4
certifi==2019.9.11
Click==7.0
configparser==4.0.2
db==0.1.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.2
marshmallow-sqlalchemy==0.19.0
mysql-connector==2.2.9
mysqlclient==1.4.5
protobuf==3.10.0
py-healthcheck==1.9.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
sentry-sdk==0.13.1
six==1.13.0
SQLAlchemy==1.3.10
urllib3==1.25.6
Werkzeug==0.16.0
